+++ src/door.js
@@ -1,4 +1,4 @@
-// Code version 15/10/2024
+// Part of garagedoor-accfactory
 // Mark Hulskamp
 'use strict';
 
@@ -13,7 +13,27 @@ import { setTimeout } from 'node:timers';
 import HomeKitDevice from './HomeKitDevice.js';
 
 export default class GarageDoor extends HomeKitDevice {
+  static TYPE = 'GarageDoor';
+  static VERSION = '2025.06.15';
+
   static DOOREVENT = 'DOOREVENT'; // Door status event tag
+
+  // Define door states
+  static OPEN = 'open';
+  static OPENED = 'opened';
+  static CLOSE = 'close';
+  static CLOSED = 'closed';
+  static OPENING = 'opening';
+  static CLOSING = 'closing';
+  static STOPPED = 'stopped';
+  static MOVING = 'moving';
+  static OBSTRUCTION = 'obstruction';
+  static UNKNOWN = 'unknown';
+  static FAULT = 'fault';
+
+  // GPIO pin min/max
+  static MIN_GPIO_PIN = 0;
+  static MAX_GPIO_PIN = 26;
 
   doorService = undefined; // HomeKit service for this garage door
   currentDoorStatus = undefined;
@@ -35,78 +55,69 @@ export default class GarageDoor extends HomeKitDevice {
       this.#eventEmitter = eventEmitter;
     }
 
-    this.currentDoorStatus = 'stopped';
-    this.#lastDoorStatus = 'unknown';
+    this.currentDoorStatus = GarageDoor.STOPPED;
+    this.#lastDoorStatus = GarageDoor.UNKNOWN;
   }
 
   // Class functions
-  addServices() {
-    // Create extra details for output
-    let postSetupDetails = [];
-
+  setupDevice() {
     // Setup the garagedoor service if not already present on the accessory
-    this.doorService = this.accessory.getService(this.hap.Service.GarageDoorOpener);
-    if (this.doorService === undefined) {
-      this.doorService = this.accessory.addService(this.hap.Service.GarageDoorOpener, '', 1);
-    }
-    if (this.doorService.testCharacteristic(this.hap.Characteristic.StatusFault) === false) {
-      // Used if the sensors report incorrect readings, such as both "high"
-      this.doorService.addCharacteristic(this.hap.Characteristic.StatusFault);
-    }
+    this.doorService = this.addHKService(this.hap.Service.GarageDoorOpener, '', 1);
     this.doorService.setPrimaryService();
+    this.addHKCharacteristic(this.doorService, this.hap.Characteristic.StatusFault);
 
     // Setup intial characteristic values
     this.doorService.updateCharacteristic(this.hap.Characteristic.StatusFault, this.hap.Characteristic.StatusFault.NO_FAULT);
     this.doorService.updateCharacteristic(this.hap.Characteristic.CurrentDoorState, this.hap.Characteristic.CurrentDoorState.STOPPED);
 
     // Setup GPIO pins
-    if (this.deviceData.pushButton === undefined || (this.deviceData.pushButton < 0 && this.deviceData.pushButton > 26)) {
-      this?.log?.warn && this.log.warn('No relay valid pin specifed for door open/close button on "%s"', this.deviceData.description);
-      this?.log?.warn && this.log.warn('We will be unable to operated garage door');
+    if (
+      isNaN(this.deviceData?.pushButton) === true ||
+      (Number(this.deviceData.pushButton) < GarageDoor.MIN_GPIO_PIN && Number(this.deviceData.pushButton > GarageDoor.MAX_GPIO_PIN))
+    ) {
+      this?.log?.warn?.('No relay valid pin specifed for door open/close button on "%s"', this.deviceData.description);
+      this?.log?.warn?.('We will be unable to operate garage door');
     }
 
-    if (this.deviceData.pushButton !== undefined && this.deviceData.pushButton >= 0 && this.deviceData.pushButton <= 26) {
+    if (this.#validGPIOPin(this.deviceData?.pushButton) === true) {
       // Push button
       GPIO.open(this.deviceData.pushButton, GPIO.OUTPUT, GPIO.LOW);
-      this?.log?.debug &&
-        this.log.debug('Setup open/close relay on "%s" using GPIO pin "%s"', this.deviceData.description, this.deviceData.pushButton);
+      this?.log?.debug?.('Setup open/close relay on "%s" using GPIO pin "%s"', this.deviceData.description, this.deviceData.pushButton);
     }
 
-    if (this.deviceData.closedSensor !== undefined && this.deviceData.closedSensor >= 0 && this.deviceData.closedSensor <= 26) {
+    if (this.#validGPIOPin(this.deviceData?.closedSensor) === true) {
       // Door closed sensor
       GPIO.open(this.deviceData.closedSensor, GPIO.INPUT, GPIO.PULL_DOWN);
-      postSetupDetails.push('Door closed sensor');
-      this?.log?.debug &&
-        this.log.debug('Setup closed door sensor on "%s" using GPIO pin "%s"', this.deviceData.description, this.deviceData.closedSensor);
-    }
-    if (this.deviceData.openSensor !== undefined && this.deviceData.openSensor >= 0 && this.deviceData.openSensor <= 26) {
-      // Door open sensor
-      GPIO.open(this.deviceData.openSensor, GPIO.INPUT, GPIO.PULL_DOWN);
-      postSetupDetails.push('Door open sensor');
-      this?.log?.debug &&
-        this.log.debug('Setup open door sensor on "%s" using GPIO pin "%s"', this.deviceData.description, this.deviceData.openSensor);
+      this.postSetupDetail('Door closed sensor');
+      this?.log?.debug?.('Setup closed door sensor on "%s" using GPIO pin "%s"', this.deviceData.description, this.deviceData.closedSensor);
     }
 
-    if (
-      this.deviceData.obstructionSensor !== undefined &&
-      this.deviceData.obstructionSensor >= 0 &&
-      this.deviceData.obstructionSensor <= 26
-    ) {
+    if (this.#validGPIOPin(this.deviceData?.openSensor) === true) {
+      // Door open sensor
+      GPIO.open(this.deviceData.openSensor, GPIO.INPUT, GPIO.PULL_DOWN);
+      this.postSetupDetail('Door open sensor');
+      this?.log?.debug?.('Setup open door sensor on "%s" using GPIO pin "%s"', this.deviceData.description, this.deviceData.openSensor);
+    }
+
+    if (this.#validGPIOPin(this.deviceData?.obstructionSensor) === true) {
       // Door obstruction sensor
       GPIO.open(this.deviceData.obstructionSensor, GPIO.INPUT, GPIO.PULL_DOWN);
-      postSetupDetails.push('Obstruction sensor');
-      this?.log?.debug &&
-        this.log.debug('Setup obstruction sensor on "%s" using GPIO pin "%s"', this.deviceData.description, this.deviceData.openSensor);
+      this.postSetupDetail('Obstruction sensor');
+      this?.log?.debug?.('Setup obstruction sensor on "%s" using GPIO pin "%s"', this.deviceData.description, this.deviceData.openSensor);
     }
 
     // Setup callbacks for characteristics
-    this.doorService.getCharacteristic(this.hap.Characteristic.TargetDoorState).onSet((value) => {
-      this.setDoorPosition(value);
+    this.addHKCharacteristic(this.doorService, this.hap.Characteristic.TargetDoorState, {
+      onSet: (value) => {
+        this.setDoorPosition(value);
+      },
     });
-    this.doorService.getCharacteristic(this.hap.Characteristic.CurrentDoorState).onGet(() => {
-      let status = this.getDoorPosition();
-      // Convert our internal string status into the HomeKit number value
-      return this.hap.Characteristic.CurrentDoorState[status.toUpperCase()];
+    this.addHKCharacteristic(this.doorService, this.hap.Characteristic.CurrentDoorState, {
+      onGet: () => {
+        let status = this.getDoorPosition();
+        // Convert our internal string status into the HomeKit number value
+        return this.hap.Characteristic.CurrentDoorState[status.toUpperCase()];
+      },
     });
 
     // Setup linkage to EveHome app if configured todo so
@@ -119,27 +130,27 @@ export default class GarageDoor extends HomeKitDevice {
         description: this.deviceData.description,
       });
     }
-
-    return postSetupDetails;
   }
 
   setDoorPosition(value) {
     // Set position of the door. (will either be open or closed)
-    if ((value === this.hap.Characteristic.TargetDoorState.CLOSED || value === 'close') && this.isClosed() === false) {
-      if (this.currentDoorState === 'opening') {
+    if ((value === this.hap.Characteristic.TargetDoorState.CLOSED || value === GarageDoor.CLOSE) && this.isClosed() === false) {
+      if (this.currentDoorStatus === GarageDoor.OPENING) {
         // Since door is "moving", press button to stop. Second press below will close ie: reverse
         this.pressButton();
-        this.#lastDoorStatus = 'stopped';
+        this.currentDoorStatus = GarageDoor.STOPPED;
+        this.#lastDoorStatus = GarageDoor.OPENING;
       }
       // "Press" garage opener/closer button, and update HomeKit status to show door moving.
       // the poll function will update to the closed status when sensor triggered
       this.pressButton();
     }
-    if ((value === this.hap.Characteristic.TargetDoorState.OPEN || value === 'open') && this.isOpen() === false) {
-      if (this.currentDoorState === 'closing') {
+    if ((value === this.hap.Characteristic.TargetDoorState.OPEN || value === GarageDoor.OPEN) && this.isOpen() === false) {
+      if (this.currentDoorStatus === GarageDoor.CLOSING) {
         // Since door is "moving", press button to stop. Second press below will close ie: reverse
         this.pressButton();
-        this.#lastDoorStatus = 'stopped';
+        this.currentDoorStatus = GarageDoor.STOPPED;
+        this.#lastDoorStatus = GarageDoor.CLOSING;
       }
       // "Press" garage opener/closer button, and update HomeKit status to show door moving.
       // the poll function will update to the open status when sensor triggered
@@ -152,7 +163,7 @@ export default class GarageDoor extends HomeKitDevice {
   }
 
   pressButton() {
-    if (this.deviceData.pushButton === undefined) {
+    if (this.#validGPIOPin(this.deviceData?.pushButton) === false) {
       return;
     }
 
@@ -162,49 +173,42 @@ export default class GarageDoor extends HomeKitDevice {
     GPIO.msleep(500);
     GPIO.write(this.deviceData.pushButton, GPIO.LOW);
     GPIO.msleep(500);
+
+    this?.log?.debug?.('Button pressed for Door "%s"', this.deviceData.description);
   }
 
-  isOpen(openSensor) {
-    if (this.deviceData.openSensor === undefined && openSensor === undefined) {
+  isOpen() {
+    if (this.#validGPIOPin(this.deviceData?.openSensor) === false) {
       return;
     }
 
-    if (openSensor === undefined && this.deviceData.openSensor !== undefined) {
-      openSensor = this.deviceData.openSensor;
-    }
-    return GPIO.read(openSensor) === GPIO.HIGH ? true : false; // If high on sensor, means door is opened
+    return GPIO.read(this.deviceData.openSensor) === GPIO.HIGH ? true : false; // If high on sensor, means door is opened
   }
 
-  isClosed(closedSensor) {
-    if (this.deviceData.closedSensor === undefined && closedSensor === undefined) {
+  isClosed() {
+    if (this.#validGPIOPin(this.deviceData?.closedSensor) === false) {
       return;
     }
 
-    if (closedSensor === undefined && this.deviceData.closedSensor !== undefined) {
-      closedSensor = this.deviceData.closedSensor;
-    }
-    return GPIO.read(closedSensor) === GPIO.HIGH ? true : false; // If high on sensor, means door is closed
+    return GPIO.read(this.deviceData.closedSensor) === GPIO.HIGH ? true : false; // If high on sensor, means door is closed
   }
 
-  hasObstruction(obstructionSensor) {
-    if (this.deviceData.obstructionSensor === undefined && obstructionSensor === undefined) {
+  hasObstruction() {
+    if (this.#validGPIOPin(this.deviceData?.obstructionSensor) === false) {
       return;
     }
 
-    if (obstructionSensor === undefined && this.deviceData.obstructionSensor !== undefined) {
-      obstructionSensor = this.deviceData.obstructionSensor;
-    }
-    return GPIO.read(obstructionSensor) === GPIO.HIGH ? true : false; // If high, obstruction detected
+    return GPIO.read(this.deviceData.obstructionSensor) === GPIO.HIGH ? true : false; // If high, obstruction detected
   }
 
-  messageServices(type, message) {
-    if (type === GarageDoor.DOOREVENT) {
-      if (message.status === 'closed' && this.currentDoorStatus !== 'closed') {
+  messageDevice(type, message) {
+    if (type === GarageDoor.DOOREVENT && typeof message?.status === 'string') {
+      if (message.status === GarageDoor.CLOSED && this.currentDoorStatus !== GarageDoor.CLOSED) {
         // Closed
         this.doorService.updateCharacteristic(this.hap.Characteristic.StatusFault, this.hap.Characteristic.StatusFault.NO_FAULT);
         this.doorService.updateCharacteristic(this.hap.Characteristic.CurrentDoorState, this.hap.Characteristic.CurrentDoorState.CLOSED);
         this.doorService.updateCharacteristic(this.hap.Characteristic.TargetDoorState, this.hap.Characteristic.CurrentDoorState.CLOSED);
-        this.currentDoorStatus = 'closed';
+        this.currentDoorStatus = GarageDoor.CLOSED;
 
         if (typeof this.historyService?.addHistory === 'function' && this.doorService !== undefined) {
           // Log door closed to history service if present
@@ -213,15 +217,15 @@ export default class GarageDoor extends HomeKitDevice {
             this.historyService.addHistory(this.doorService, { time: Math.floor(Date.now() / 1000), status: 0 }); // closed
           }
         }
-        this?.log?.success && this.log.success('Door "%s" is closed', this.deviceData.description);
+        this?.log?.success?.('Door "%s" is closed', this.deviceData.description);
       }
 
-      if (message.status === 'open' && this.currentDoorStatus !== 'open') {
+      if (message.status === GarageDoor.OPENED && this.currentDoorStatus !== GarageDoor.OPENED) {
         // Open
         this.doorService.updateCharacteristic(this.hap.Characteristic.StatusFault, this.hap.Characteristic.StatusFault.NO_FAULT);
         this.doorService.updateCharacteristic(this.hap.Characteristic.CurrentDoorState, this.hap.Characteristic.CurrentDoorState.OPEN);
         this.doorService.updateCharacteristic(this.hap.Characteristic.TargetDoorState, this.hap.Characteristic.CurrentDoorState.OPEN);
-        this.currentDoorStatus = 'open';
+        this.currentDoorStatus = GarageDoor.OPENED;
 
         if (typeof this.historyService?.addHistory === 'function' && this.doorService !== undefined) {
           // Log door opened to history service if present
@@ -230,35 +234,35 @@ export default class GarageDoor extends HomeKitDevice {
             this.historyService.addHistory(this.doorService, { time: Math.floor(Date.now() / 1000), status: 1 }); // open
           }
         }
-        this?.log?.warn && this.log.warn('Door "%s" is open', this.deviceData.description);
+        this?.log?.warn?.('Door "%s" is open', this.deviceData.description);
       }
 
-      if (message.status === 'moving') {
+      if (message.status === GarageDoor.MOVING) {
         // Moving
         this.doorService.updateCharacteristic(this.hap.Characteristic.StatusFault, this.hap.Characteristic.StatusFault.NO_FAULT);
-        if (message.last === 'closed' && this.currentDoorStatus !== 'opening') {
+        if (message.last === GarageDoor.CLOSED && this.currentDoorStatus !== GarageDoor.OPENING) {
           // Since door was last closed, and now its moving, assume its opening
           this.doorService.updateCharacteristic(this.hap.Characteristic.CurrentDoorState, this.hap.Characteristic.CurrentDoorState.OPENING);
           this.doorService.updateCharacteristic(this.hap.Characteristic.TargetDoorState, this.hap.Characteristic.CurrentDoorState.OPEN);
-          this.currentDoorStatus = 'opening';
-          this?.log?.debug && this.log.debug('Door "%s" is opening', this.deviceData.description);
+          this.currentDoorStatus = GarageDoor.OPENING;
+          this?.log?.debug?.('Door "%s" is opening', this.deviceData.description);
         }
-        if (message.last === 'open' && this.currentDoorStatus !== 'closing') {
+        if (message.last === GarageDoor.OPENED && this.currentDoorStatus !== GarageDoor.CLOSING) {
           // Since door was last open, and now its moving, assume its closing
           this.doorService.updateCharacteristic(this.hap.Characteristic.CurrentDoorState, this.hap.Characteristic.CurrentDoorState.CLOSING);
           this.doorService.updateCharacteristic(this.hap.Characteristic.TargetDoorState, this.hap.Characteristic.CurrentDoorState.CLOSED);
-          this.currentDoorStatus = 'closing';
-          this?.log?.debug && this.log.debug('Door "%s" is closing', this.deviceData.description);
+          this.currentDoorStatus = GarageDoor.CLOSING;
+          this?.log?.debug?.('Door "%s" is closing', this.deviceData.description);
         }
       }
 
-      if (message.status === 'stopped') {
+      if (message.status === GarageDoor.STOPPED) {
         // Stopped
-        if (this.currentDoorStatus !== 'stopped') {
+        if (this.currentDoorStatus !== GarageDoor.STOPPED) {
           this.doorService.updateCharacteristic(this.hap.Characteristic.StatusFault, this.hap.Characteristic.StatusFault.NO_FAULT);
           this.doorService.updateCharacteristic(this.hap.Characteristic.CurrentDoorState, this.hap.Characteristic.CurrentDoorState.STOPPED);
           this.doorService.updateCharacteristic(this.hap.Characteristic.TargetDoorState, this.hap.Characteristic.CurrentDoorState.OPEN);
-          this.currentDoorStatus = 'stopped';
+          this.currentDoorStatus = GarageDoor.STOPPED;
 
           if (typeof this.historyService?.addHistory === 'function' && this.doorService !== undefined) {
             // Log door opened to history service if present
@@ -267,31 +271,34 @@ export default class GarageDoor extends HomeKitDevice {
               this.historyService.addHistory(this.doorService, { time: Math.floor(Date.now() / 1000), status: 1 }); // open
             }
           }
-          this?.log?.debug && this.log.debug('Door "%s" has stopped moving', this.deviceData.description);
+          this?.log?.debug?.('Door "%s" has stopped moving', this.deviceData.description);
         }
       }
 
-      if (message.status === 'fault') {
+      if (message.status === GarageDoor.FAULT) {
         // Faulty sensors
         this.doorService.updateCharacteristic(this.hap.Characteristic.StatusFault, this.hap.Characteristic.StatusFault.GENERAL_FAULT);
         this.doorService.updateCharacteristic(this.hap.Characteristic.CurrentDoorState, this.hap.Characteristic.CurrentDoorState.STOPPED);
         // What should current door status be???
-        this?.log?.error && this.log.error('Door "%s" is reporting fault with sensors', this.deviceData.description);
+        this?.log?.error?.('Door "%s" is reporting fault with sensors', this.deviceData.description);
       }
 
-      if (message.status === 'obstruction' || message.status === 'clear') {
+      if (message.status === GarageDoor.OBSTRUCTION || message.status === 'clear') {
         // Door obstruction, either active or cleared
-        this.doorService.updateCharacteristic(this.hap.Characteristic.ObstructionDetected, message.status === 'obstruction' ? true : false);
-        this?.log?.warn && this.log.warn('Door "%s" is reporting an obstruction', this.deviceData.description);
+        this.doorService.updateCharacteristic(
+          this.hap.Characteristic.ObstructionDetected,
+          message.status === GarageDoor.OBSTRUCTION ? true : false,
+        );
+        this?.log?.warn?.('Door "%s" is reporting an obstruction', this.deviceData.description);
         // <---- Implement. Do we stop door from being closed if obstructed? or just allow to open?? or not allow movement at all?
       }
     }
   }
 
-  updateServices(deviceData) {
-    let doorOpen = this.isOpen(deviceData.openSensor);
-    let doorClosed = this.isClosed(deviceData.closeSensor);
-    let obstruction = this.hasObstruction(deviceData.obstructionSensor);
+  updateDevice(deviceData) {
+    let doorOpen = this.isOpen();
+    let doorClosed = this.isClosed();
+    let obstruction = this.hasObstruction();
 
     // Work out the current status of the door using configured sensors.
     // This will either be "open", "closed", "moving", "stopped"
@@ -299,73 +306,77 @@ export default class GarageDoor extends HomeKitDevice {
     if (doorClosed === true && doorOpen === false) {
       // Door is fully closed
       this.#moveStartedTime = undefined;
-      this.#lastDoorStatus = 'closed';
-      if (this.#eventEmitter !== undefined) {
-        this.#eventEmitter.emit(this.uuid, GarageDoor.DOOREVENT, { status: 'closed' });
-      }
+      this.#lastDoorStatus = GarageDoor.CLOSED;
+      this.#eventEmitter?.emit?.(this.uuid, GarageDoor.DOOREVENT, { status: GarageDoor.CLOSED });
     }
+
     if (doorClosed === false && doorOpen === true) {
       // Door is fully open
       this.#moveStartedTime = undefined;
-      this.#lastDoorStatus = 'open';
-      if (this.#eventEmitter !== undefined) {
-        this.#eventEmitter.emit(this.uuid, GarageDoor.DOOREVENT, { status: 'open' });
-      }
+      this.#lastDoorStatus = GarageDoor.OPENED;
+      this.#eventEmitter?.emit?.(this.uuid, GarageDoor.DOOREVENT, { status: GarageDoor.OPENED });
     }
+
     if (doorClosed === false && doorOpen === false) {
       // Door is neither open or closed, so door is either moving or stopped.
       if (this.#moveStartedTime === undefined) {
         this.#moveStartedTime = Date.now(); // Time we detected first movement from either open or closed
 
-        if (this.#lastDoorStatus === 'stopped') {
+        if (this.#lastDoorStatus === GarageDoor.STOPPED) {
           // Detected movement after stopped state, ie: we've pressed the push button
           // Stopped state is assumed to be door open, as neither detected open or closed
-          this.#lastDoorStatus = 'open';
+          this.#lastDoorStatus = this.currentDoorStatus;
         }
+      }
+
+      // Reset timer if direction reversed during movement
+      if (
+        (this.#lastDoorStatus === GarageDoor.OPENING && this.currentDoorStatus === GarageDoor.CLOSING) ||
+        (this.#lastDoorStatus === GarageDoor.CLOSING && this.currentDoorStatus === GarageDoor.OPENING)
+      ) {
+        this.#moveStartedTime = Date.now();
+        this.#lastDoorStatus = this.currentDoorStatus;
       }
 
       let duration = Math.floor(Date.now()) - (this.#moveStartedTime !== undefined ? this.#moveStartedTime : 0);
       if (
-        this.#lastDoorStatus === 'unknown' ||
-        this.#lastDoorStatus === 'stopped' ||
-        (this.#lastDoorStatus === 'open' && duration > deviceData.closeTime * 1000) ||
-        (this.#lastDoorStatus === 'closed' && duration > deviceData.openTime * 1000)
+        this.#lastDoorStatus === GarageDoor.UNKNOWN ||
+        this.#lastDoorStatus === GarageDoor.STOPPED ||
+        (this.#lastDoorStatus === GarageDoor.OPENED && duration > deviceData.closeTime * 1000) ||
+        (this.#lastDoorStatus === GarageDoor.CLOSED && duration > deviceData.openTime * 1000)
       ) {
         // Since the door state isn't open or closed OR open or closed status and moving time has been exceeded for configured times
         // In this case we'll assume door has stopped
-        this.#lastDoorStatus = 'stopped';
-        if (this.#eventEmitter !== undefined) {
-          this.#eventEmitter.emit(this.uuid, GarageDoor.DOOREVENT, { status: 'stopped' });
-        }
+        this.#lastDoorStatus = GarageDoor.STOPPED;
+        this.#eventEmitter?.emit?.(this.uuid, GarageDoor.DOOREVENT, { status: GarageDoor.STOPPED });
       } else {
-        if (this.#eventEmitter !== undefined) {
-          this.#eventEmitter.emit(this.uuid, GarageDoor.DOOREVENT, {
-            status: 'moving',
-            last: this.#lastDoorStatus,
-            duration: duration,
-          });
-        }
+        this.#eventEmitter?.emit?.(this.uuid, GarageDoor.DOOREVENT, {
+          status: GarageDoor.MOVING,
+          last: this.#lastDoorStatus,
+          duration: duration,
+        });
       }
     }
-    if (doorClosed === true && doorOpen === true) {
-      // Is reading both open and close, we'll assume fault with sensors
-      if (this.#eventEmitter !== undefined) {
-        this.#eventEmitter.emit(this.uuid, GarageDoor.DOOREVENT, { status: 'fault', last: this.#lastDoorStatus });
-      }
+
+    if ((doorClosed === true && doorOpen === true) || doorClosed === undefined || doorOpen === undefined) {
+      // Is reading both open and close OR no status, we'll assume fault with sensors
+      this.#eventEmitter?.emit?.(this.uuid, GarageDoor.DOOREVENT, { status: GarageDoor.FAULT, last: this.#lastDoorStatus });
     }
 
     if (obstruction !== undefined) {
       // Since obstruction didn't return an undefined value, this means we have a configured obstruction sensor and its returned its status
-      if (this.#eventEmitter !== undefined) {
-        this.#eventEmitter.emit(this.uuid, GarageDoor.DOOREVENT, { status: obstruction === true ? 'obstruction' : 'clear' });
-      }
+      this.#eventEmitter?.emit?.(this.uuid, GarageDoor.DOOREVENT, { status: obstruction === true ? GarageDoor.OBSTRUCTION : 'clear' });
     }
 
     // Perform this again after a short period by issuing an device update message
-    // The updateServices function will only be called again from this message if some data has changed
+    // The updateDevice function will only be called again from this message if some data has changed
     // We can force this by adding a "timestamp" field to the data object
     setTimeout(() => {
-      this.#eventEmitter.emit(this.uuid, HomeKitDevice.UPDATE, { lastDoorCheckTime: Date.now() });
+      this.#eventEmitter?.emit?.(this.uuid, HomeKitDevice.UPDATE, { lastDoorCheckTime: Date.now() });
     }, 1000);
+  }
+
+  #validGPIOPin(pin) {
+    return isNaN(pin) === false && Number(pin) >= GarageDoor.MIN_GPIO_PIN && Number(pin) <= GarageDoor.MAX_GPIO_PIN;
   }
 }